fişx

+++ rife/warplayer.py
@@ -4,8 +4,11 @@ import torch.nn as nn
 device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 backwarp_tenGrid = {}
 
-
 def warp(tenInput, tenFlow):
+    # Ensure input tensors are BFloat16
+    tenInput = tenInput.to(dtype=torch.bfloat16)
+    tenFlow = tenFlow.to(dtype=torch.bfloat16)
+
     k = (str(tenFlow.device), str(tenFlow.size()))
     if k not in backwarp_tenGrid:
         tenHorizontal = (
@@ -18,7 +21,7 @@ def warp(tenInput, tenFlow):
             .view(1, 1, tenFlow.shape[2], 1)
             .expand(tenFlow.shape[0], -1, -1, tenFlow.shape[3])
         )
-        backwarp_tenGrid[k] = torch.cat([tenHorizontal, tenVertical], 1).to(device)
+        backwarp_tenGrid[k] = torch.cat([tenHorizontal, tenVertical], 1).to(device).to(dtype=torch.bfloat16)
 
     tenFlow = torch.cat(
         [
@@ -30,5 +33,9 @@ def warp(tenInput, tenFlow):
 
     g = (backwarp_tenGrid[k] + tenFlow).permute(0, 2, 3, 1)
     return torch.nn.functional.grid_sample(
-        input=tenInput, grid=g, mode="bilinear", padding_mode="border", align_corners=True
-    )
+        input=tenInput, 
+        grid=g, 
+        mode="bilinear", 
+        padding_mode="border", 
+        align_corners=True
+    ).to(dtype=torch.bfloat16)  # Ensure output is BFloat16
